// src/types/Product-type.ts
export interface CardNewModelsType {
    id: number;
    category: string;
    itemId: string;
    name: string;
    fullPrice: number;
    price: number;
    screen: string,
    capacity: string,
    color: string,
    ram: string,
    year: number,
    image: string,
}

export type CardNewModels = CardNewModelsType[];